@@ -14,10 +14,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import { TestBed } from '@angular/core/testing';
-import {
-  HttpClientTestingModule,
-  HttpTestingController,
-} from '@angular/common/http/testing';
+import { HttpTestingController, provideHttpClientTesting } from '@angular/common/http/testing';
 import { CRUDService } from './crud.service';
 import {
   CatalogService,
@@ -26,6 +23,7 @@ import {
   Catalog,
 } from './catalog.service';
 import { AuthService } from './auth.service';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('CatalogService', () => {
   let sut: CatalogService;
@@ -36,8 +34,9 @@ describe('CatalogService', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
-    });
+    imports: [],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+});
     TestBed.inject(AuthService).setAccessToken({
       access_token: 'token',
       token_type: 'bearer',
