@@ -14,16 +14,14 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import { TestBed } from '@angular/core/testing';
-import {
-  HttpClientTestingModule,
-  HttpTestingController,
-} from '@angular/common/http/testing';
+import { HttpTestingController, provideHttpClientTesting } from '@angular/common/http/testing';
 import { CRUDService } from './crud.service';
 import { AuthService } from './auth.service';
 import {
   IJiraIssueType,
   JiraIssueTypeService,
 } from './jira-issue-type.service';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('JiraIssueTypeService', () => {
   let sut: JiraIssueTypeService;
@@ -33,8 +31,9 @@ describe('JiraIssueTypeService', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
-    });
+    imports: [],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+});
     TestBed.inject(AuthService).setAccessToken({
       access_token: 'token',
       token_type: 'bearer',
