@@ -14,10 +14,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import { TestBed } from '@angular/core/testing';
-import {
-  HttpClientTestingModule,
-  HttpTestingController,
-} from '@angular/common/http/testing';
+import { HttpTestingController, provideHttpClientTesting } from '@angular/common/http/testing';
 import { CRUDService } from './crud.service';
 import { AuthService } from './auth.service';
 import {
@@ -25,6 +22,7 @@ import {
   IJiraIssue,
   JiraIssueService,
 } from './jira-issue.service';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('JiraIssueService', () => {
   let sut: JiraIssueService;
@@ -35,8 +33,9 @@ describe('JiraIssueService', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
-    });
+    imports: [],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+});
     TestBed.inject(AuthService).setAccessToken({
       access_token: 'token',
       token_type: 'bearer',
